Starting Part 2 solution - end of time available
Added the fundamental changes to the functions and classes from part one to accommodate solution in part two. The SchematicPart got a new field to store the location of the associated gear. The process_schematic function now accepts the x,y co-ords of the source symbol and if it was a gear.